Feat: clean up types
